+++ package.json
@@ -2,9 +2,6 @@
 	"name": "quantacms",
 	"version": "0.0.1",
 	"private": true,
-	"prisma": {
-		"seed": "ts-node prisma/seed.js"
-	},
 	"scripts": {
 		"dev": "vite dev --host",
 		"build": "vite build",
@@ -19,7 +16,6 @@
 		"@sveltejs/adapter-auto": "^1.0.0",
 		"@sveltejs/adapter-node": "^1.1.4",
 		"@sveltejs/kit": "^1.0.0",
-		"@types/node": "^18.13.0",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
 		"@typescript-eslint/parser": "^5.45.0",
 		"autoprefixer": "^10.4.13",
@@ -36,9 +32,8 @@
 		"svelte": "^3.54.0",
 		"svelte-check": "^3.0.1",
 		"tailwindcss": "^3.2.4",
-		"ts-node": "^10.9.1",
 		"tslib": "^2.4.1",
-		"typescript": "^4.9.5",
+		"typescript": "^4.9.3",
 		"vite": "^4.0.0"
 	},
 	"type": "module",
