+++ prisma/schema.prisma
@@ -12,22 +12,22 @@ datasource db {
 
 model SalesRep {
   uid String @id @default(cuid())
-  name String @unique
+  name String
   comments String
   clients Client[]
 }
 
 model Client {
   uid String @id @default(cuid())
-  name String @unique
+  name String
   salesRep SalesRep @relation(fields: [salesRepId], references: [uid])
   salesRepId String
   locations Location[]
+  contracts Contract[]
 }
 
 model Location {
   uid String @id @default(cuid())
-  name String @unique
   address String
   client Client @relation(fields: [clientId], references: [uid])
   clientId String
@@ -37,11 +37,10 @@ model Location {
 model Device {
   uid String @id @default(cuid())
   location Location @relation(fields: [locationId], references: [uid])
-  contract Contract @relation(fields: [contractId], references: [uid])
   locationId String
-  contractId String
-  serialNum String @unique
-  modelNum String
+  serialNum Int
+  partNum String
+  invoice Int
   installed DateTime
   deviceEvents DeviceEvent[]
   repairEvents RepairEvent[]
@@ -58,12 +57,13 @@ model DeviceEvent {
 
 model Contract {
   uid String @id @default(cuid())
+  client Client @relation(fields: [clientId], references: [uid])
+  clientId String
   start DateTime
   end DateTime
   type String
   comments String
   contractEvents ContractEvent[]
-  devices Device[]
 }
 
 model ContractEvent {
@@ -76,7 +76,7 @@ model ContractEvent {
 
 model ServiceRep {
   uid String @id @default(cuid())
-  name String @unique
+  name String
   email String
   comments String
   serviceCalls ServiceCall[]
