--- src/lib/ToastPopup.svelte
@@ -1,38 +1,0 @@
-<script lang="ts">
-	import { Toast } from 'flowbite-svelte';
-	import { slide } from 'svelte/transition';
-
-	export let show: boolean;
-	export let message: string;
-
-	$: {
-		if (show) {
-			setTimeout(() => {
-				show = false;
-			}, 5000);
-		}
-	}
-</script>
-
-<div class="flex justify-center">
-	<div class="mb-2 mt-5 absolute">
-		<Toast color="pink" simple transition={slide} bind:open={show}>
-			<svelte:fragment slot="icon">
-				<svg
-					aria-hidden="true"
-					class="w-5 h-5"
-					fill="currentColor"
-					viewBox="0 0 20 20"
-					xmlns="http://www.w3.org/2000/svg"
-					><path
-						fill-rule="evenodd"
-						d="M8.257 3.099c.765-1.36 2.722-1.36 3.486 0l5.58 9.92c.75 1.334-.213 2.98-1.742 2.98H4.42c-1.53 0-2.493-1.646-1.743-2.98l5.58-9.92zM11 13a1 1 0 11-2 0 1 1 0 012 0zm-1-8a1 1 0 00-1 1v3a1 1 0 002 0V6a1 1 0 00-1-1z"
-						clip-rule="evenodd"
-					/></svg
-				>
-				<span class="sr-only">Warning icon</span>
-			</svelte:fragment>
-			{message}
-		</Toast>
-	</div>
-</div>
