--- src/lib/database.ts
@@ -1,3 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-export const prisma = new PrismaClient();
