+++ src/routes/client/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Test!</h1>
